evaluation/requirements.txt
azure-ai-evaluation>=1.13.7
httpx>=0.28.1
